Fix a shitton of runtimes when the reactor detonates

+++ code/modules/power/rbmk/decals.dm
@@ -9,6 +9,9 @@
 
 /obj/effect/decal/nuclear_waste/Initialize()
 	. = ..()
+	if(. == INITIALIZE_HINT_QDEL)
+		return
+
 	for(var/obj/A in get_turf(src))
 		if(istype(A, /obj/structure))
 			qdel(src) //It is more processing efficient to do this here rather than when searching for available turfs.
@@ -41,11 +44,11 @@
 	if(isliving(AM))
 		var/mob/living/L = AM
 		playsound(loc, 'sound/effects/gib_step.ogg', HAS_TRAIT(L, TRAIT_LIGHT_STEP) ? 20 : 50, 1)
-	radiation_pulse(src, 500, 5) //MORE RADS
+	radiation_pulse(src, 2, RAD_LIGHT_INSULATION, IRRADIATION_CHANCE_DEPLETED_FUEL) //MORE RADS
 
 /obj/effect/decal/nuclear_waste/attackby(obj/item/tool, mob/user)
 	if(tool.tool_behaviour == TOOL_SHOVEL)
-		radiation_pulse(src, 1000, 5) //MORE RADS
+		radiation_pulse(src, 2, RAD_LIGHT_INSULATION, IRRADIATION_CHANCE_DEPLETED_FUEL) //MORE RADS
 		to_chat(user, "<span class='notice'>You start to clear [src]...</span>")
 		if(tool.use_tool(src, user, 50, volume=100))
 			to_chat(user, "<span class='notice'>You clear [src]. </span>")
